Removed 'amplify' and 'be' folders from branch, moved comparator data structure to appropriate folder, added basic authentication (Google OAuth and AWS Amplify)
