--- amplify/.config/project-config.json
@@ -1,17 +1,0 @@
-{
-  "projectName": "JungleCart",
-  "version": "3.1",
-  "frontend": "javascript",
-  "javascript": {
-    "framework": "none",
-    "config": {
-      "SourceDir": "src",
-      "DistributionDir": "dist",
-      "BuildCommand": "npm run-script build",
-      "StartCommand": "npm run-script start"
-    }
-  },
-  "providers": [
-    "awscloudformation"
-  ]
-}
