--- amplify/backend/auth/junglecarte75abb71/cli-inputs.json
@@ -1,46 +1,0 @@
-{
-  "version": "1",
-  "cognitoConfig": {
-    "identityPoolName": "junglecarte75abb71_identitypool_e75abb71",
-    "allowUnauthenticatedIdentities": false,
-    "resourceNameTruncated": "junglee75abb71",
-    "userPoolName": "junglecarte75abb71_userpool_e75abb71",
-    "autoVerifiedAttributes": [
-      "email"
-    ],
-    "mfaConfiguration": "OFF",
-    "mfaTypes": [
-      "SMS Text Message"
-    ],
-    "smsAuthenticationMessage": "Your authentication code is {####}",
-    "smsVerificationMessage": "Your verification code is {####}",
-    "emailVerificationSubject": "Your verification code",
-    "emailVerificationMessage": "Your verification code is {####}",
-    "defaultPasswordPolicy": false,
-    "passwordPolicyMinLength": 8,
-    "passwordPolicyCharacters": [],
-    "requiredAttributes": [
-      "email"
-    ],
-    "aliasAttributes": [],
-    "userpoolClientGenerateSecret": false,
-    "userpoolClientRefreshTokenValidity": 30,
-    "userpoolClientWriteAttributes": [
-      "email"
-    ],
-    "userpoolClientReadAttributes": [
-      "email"
-    ],
-    "userpoolClientLambdaRole": "junglee75abb71_userpoolclient_lambda_role",
-    "userpoolClientSetAttributes": false,
-    "sharedId": "e75abb71",
-    "resourceName": "junglecarte75abb71",
-    "authSelections": "identityPoolAndUserPool",
-    "useDefault": "default",
-    "triggers": {},
-    "userPoolGroupList": [],
-    "serviceName": "Cognito",
-    "usernameCaseSensitive": false,
-    "useEnabledMfas": true
-  }
-}
