--- amplify/team-provider-info.json
@@ -1,20 +1,0 @@
-{
-  "dev": {
-    "awscloudformation": {
-      "AuthRoleName": "amplify-junglecart-dev-01054-authRole",
-      "UnauthRoleArn": "arn:aws:iam::846202041248:role/amplify-junglecart-dev-01054-unauthRole",
-      "AuthRoleArn": "arn:aws:iam::846202041248:role/amplify-junglecart-dev-01054-authRole",
-      "Region": "us-west-2",
-      "DeploymentBucketName": "amplify-junglecart-dev-01054-deployment",
-      "UnauthRoleName": "amplify-junglecart-dev-01054-unauthRole",
-      "StackName": "amplify-junglecart-dev-01054",
-      "StackId": "arn:aws:cloudformation:us-west-2:846202041248:stack/amplify-junglecart-dev-01054/1981f3b0-b5ad-11ed-ba2e-02c83a305675",
-      "AmplifyAppId": "dmjk7pwujg9af"
-    },
-    "categories": {
-      "auth": {
-        "junglecarte75abb71": {}
-      }
-    }
-  }
-}
