+++ frontend/src/pages/login/index.jsx
@@ -2,6 +2,7 @@ import { useRouter } from 'next/router';
 import Image from 'next/image';
 import styles from './login.module.css';
 import YourSvg from '../../../public/login.svg';
+import login from "./login";
 
 const Login = () => {
   const router = useRouter();
@@ -26,6 +27,7 @@ const Login = () => {
                   <button
                     className="btn btn-primary"
                     onClick={() => {
+                      login();
                       router.push('/products');
                     }}
                   >
